+++ internal/cmd/graph_generator/main.go
@@ -51,7 +51,8 @@ var (
 		"Transpose", "Where",
 
 		// Fused ops: exported wrappers with "Internal:" comments are hand-written in ops_fused.go.
-		"FusedDense", "FusedGelu", "FusedLayerNorm", "FusedSoftmax",
+		"FusedActivation", "FusedActivationVJP", "FusedDense", "FusedDenseVJP",
+		"FusedLayerNorm", "FusedSoftmax",
 		"FusedScaledDotProductAttention", "FusedScaledDotProductAttentionVJP",
 		"FusedAttentionQKVProjection",
 		"FusedQuantizedDense", "QuantizedEmbeddingLookup",
@@ -67,13 +68,15 @@ var (
 	// methodsNotGenerated get a NodeType but no auto-generated wrapper
 	// (hand-written implementations).
 	methodsNotGenerated = sets.MakeWith(
-		"Constant", "Parameter", "FusedQuantizedDense", "QuantizedEmbeddingLookup")
+		"Constant", "Parameter", "FusedQuantizedDense", "QuantizedEmbeddingLookup", "FusedDenseVJP")
 
 	// nillableParams lists Value parameters that can be nil (passed as *Node).
 	// Key format: "MethodName.paramName"
 	nillableParams = sets.MakeWith(
 		"FusedLayerNorm.gamma", "FusedLayerNorm.beta",
 		"FusedDense.bias",
+		"FusedDenseVJP.bias",
+		"FusedActivationVJP.x",
 		"FusedAttentionQKVProjection.biasQ", "FusedAttentionQKVProjection.biasK", "FusedAttentionQKVProjection.biasV",
 	)
 
@@ -180,6 +183,9 @@ func buildMethodInfo() (methods []*MethodInfo) {
 			case "ActivationType":
 				pi.BackendType = "compute." + pi.BackendType
 				pi.Format = "%s"
+			case "ActivationConfig":
+				pi.BackendType = "compute." + pi.BackendType
+				pi.Format = "%+v"
 			case "DenseConfig":
 				pi.BackendType = "compute." + pi.BackendType
 				pi.Format = "%+v"